refactor: components 폴더 분리
